pyproject.toml
[project]
name = "brickbot2026"
version = "0.1.0"
description = "Brickbot - DAIS 2026 conference assistant powered by Databricks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "mlflow>=3.10.0",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.2.1",
    "uuid-utils>=0.10.0",
    "databricks-openai[memory]>=0.13.0",
    "databricks-sdk>=0.79.0",
    "databricks-agents>=1.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_server", "scripts"]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
]

[project.scripts]
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
